[TH2-846] Move files.
