+++ src/main/kotlin/com/exactpro/th2/service/generator/protoc/Main.kt
@@ -15,10 +15,10 @@
  */
 package com.exactpro.th2.service.generator.protoc
 
-import com.exactpro.th2.service.generator.protoc.generator.java.MetaInfGenerator
-import com.exactpro.th2.service.generator.protoc.generator.java.ServiceDefaultImplGenerator
-import com.exactpro.th2.service.generator.protoc.generator.java.ServiceInterfaceGenerator
-import com.exactpro.th2.service.generator.protoc.generator.python.PythonServiceGenerator
+import com.exactpro.th2.service.generator.protoc.java.MetaInfGenerator
+import com.exactpro.th2.service.generator.protoc.java.ServiceDefaultImplGenerator
+import com.exactpro.th2.service.generator.protoc.java.ServiceInterfaceGenerator
+import com.exactpro.th2.service.generator.protoc.python.PythonServiceGenerator
 import com.exactpro.th2.service.generator.protoc.run.CommandLineRun
 import com.exactpro.th2.service.generator.protoc.run.PluginRun
