+++ src/main/kotlin/com/exactpro/th2/service/generator/protoc/java/MetaInfGenerator.kt
@@ -13,10 +13,10 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.exactpro.th2.service.generator.protoc.generator.java
+package com.exactpro.th2.service.generator.protoc.java
 
-import com.exactpro.th2.service.generator.protoc.generator.FileSpec
-import com.exactpro.th2.service.generator.protoc.generator.FileSpecImpl
+import com.exactpro.th2.service.generator.protoc.FileSpec
+import com.exactpro.th2.service.generator.protoc.FileSpecImpl
 import com.exactpro.th2.service.generator.protoc.util.javaPackage
 import com.google.protobuf.DescriptorProtos.FileDescriptorProto
 import com.google.protobuf.DescriptorProtos.ServiceDescriptorProto
